primitive: fix inconsistency between types.
	* include/urbi/object/primitive.hh (values_type): use size_t, as
	that's the type used elsewhere.
	* src/object/primitive.cc: Style change.

+++ include/urbi/object/primitive.hh
@@ -30,7 +30,7 @@ namespace urbi
     {
       public:
       typedef boost::function1<rObject, const objects_type&> value_type;
-      typedef boost::unordered_map<unsigned, value_type> values_type;
+      typedef boost::unordered_map<size_t, value_type> values_type;
 
       Primitive();
       Primitive(const value_type& p);
